+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `BiDoc` class.
 - Add `Definition.fragment`.
+- Add `out.html` option in `render` function.
 - Optimize code.
 
 ## v0.1.1